feat: adding claim command to bot engine

+++ store/interface.go
@@ -3,15 +3,15 @@ package store
 import "time"
 
 type ClaimTransaction struct {
-	TxID   string `json:"transaction_id"`
-	Data   string `json:"data"`
-	Amount int64  `json:"amount"`
-	Time   int64  `json:"time"`
+	TxID   string  `json:"transaction_id"`
+	Data   string  `json:"data"`
+	Amount float64 `json:"amount"`
+	Time   int64   `json:"time"`
 }
 
 type Claimer struct {
 	DiscordID        string            `json:"discord_id"`
-	TotalReward      int64             `json:"total_reward"`
+	TotalReward      float64           `json:"total_reward"`
 	ClaimTransaction *ClaimTransaction `json:"claim_transaction"`
 }
 
@@ -21,5 +21,5 @@ func (c *Claimer) IsClaimed() bool {
 
 type IStore interface {
 	ClaimerInfo(discordID string) *Claimer
-	AddClaimTransaction(TxID string, Amount int64, Time time.Time, Data string, discordID string) error
+	AddClaimTransaction(TxID string, Amount float64, Time time.Time, Data string, discordID string) error
 }
