+++ store/store_test.go
@@ -36,7 +36,7 @@ func TestStore(t *testing.T) {
 
 	t.Run("get claimer", func(t *testing.T) {
 		claimer := store.ClaimerInfo("123456789")
-		assert.Equal(t, int64(100), claimer.TotalReward)
+		assert.Equal(t, float64(100), claimer.TotalReward)
 		assert.Equal(t, "123456789", claimer.DiscordID)
 	})
 
@@ -57,11 +57,11 @@ func TestStore(t *testing.T) {
 		claimedInfo := store.ClaimerInfo(discordID)
 		assert.Equal(t, data, claimedInfo.ClaimTransaction.Data)
 		assert.Equal(t, discordID, claimedInfo.DiscordID)
-		assert.Equal(t, int64(100), claimedInfo.ClaimTransaction.Amount)
+		assert.Equal(t, float64(100), claimedInfo.ClaimTransaction.Amount)
 		assert.Equal(t, txID, claimedInfo.ClaimTransaction.TxID)
 		assert.Equal(t, time.Unix(), claimedInfo.ClaimTransaction.Time)
 		assert.Equal(t, claimer.TotalReward, claimedInfo.ClaimTransaction.Amount)
-		assert.Equal(t, int64(0), claimedInfo.TotalReward)
+		assert.Equal(t, float64(0), claimedInfo.TotalReward)
 
 		isClaimed = claimedInfo.IsClaimed()
 		assert.True(t, isClaimed)
